Use CommonJS file as interface between bull and job processors ES modules
